🐛 fix: 마이페이지 새로고침 시 빈데이터 받아와지는 것 수정
