+++ src/pages/Mypage/index.tsx
@@ -1,12 +1,9 @@
 import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import { useAtomValue, useSetAtom } from 'jotai';
+import { useAtomValue } from 'jotai';
 import Sidebar from '@/components/Mypage/Sidebar';
-import useUser from '@/hooks/api/useUser';
-import { ACCESS_USER_ID } from '@/constants/api';
 import { darkAtom } from '@/store/theme';
 import { userAtom } from '@/store/user';
-import { getStorage } from '@/utils/LocalStorage';
 import * as Style from './index.style';
 
 function Mypage() {
@@ -14,25 +11,14 @@ function Mypage() {
 
   const location = useLocation();
 
-  const setUser = useSetAtom(userAtom);
-
-  const userId = getStorage(ACCESS_USER_ID, '');
-
-  const { data: userData } = useUser(userId);
-
   const { pathname } = location;
 
   const darkMode = useAtomValue(darkAtom);
+  const userData = useAtomValue(userAtom);
 
   useEffect(() => {
     setIsMypage(pathname.replace('/mypage', '').length <= 1);
   }, [location, pathname]);
-
-  useEffect(() => {
-    if (userData) {
-      setUser(userData);
-    }
-  }, [setUser, userData]);
 
   return (
     <Style.Container
